+++ golem/cli/argument_parser.py
@@ -25,7 +25,9 @@ def get_parser():
 
     # gui
     parser_gui = subparsers.add_parser('gui', add_help=False)
-    parser_gui.add_argument('-p', '--port', nargs='?', default=5000, type=int)
+    parser_gui.add_argument('--host', action='store', nargs='?', default=None)
+    parser_gui.add_argument('-p', '--port', action='store', nargs='?', default=5000, type=int)
+    parser_gui.add_argument('-d', '--debug', action='store_true', default=False)
     parser_gui.add_argument('-h', '--help', action='store_true')
 
     # createproject