+++ CHANGELOG.md
@@ -13,7 +13,9 @@
     * Add and edit tags using the UI
     * Accept *tag expressions* for complex comparisons 
     * Docs: [Filter Tests by Tags](https://golem-framework.readthedocs.io/en/latest/running-tests.html#filter-tests-by-tags)
-    
+- New args to ```golem gui``` command: --host (default 127.0.0.1), -d|--debug (default False)
+- New actions: assert_element_value, assert_element_value_is_not, verify_element_value, verify_element_value_is_not, send_keys_with_delay
+- New WebElement methods: inner_html, outer_html, send_keys_with_delay
 
 ## [0.7.0] - 2019-01-21